+++ LookupAnything/release-notes.md
@@ -6,6 +6,7 @@
 * Fixed tree lookups incorrectly stating that a fertilized tree won't grow in winter.
 * Fixed error when trying to lookup on the 'help wanted' screen.
 * Fixed untranslated text in child lookup.
+* Fixed 'number owned' field not counting items in the island farmhouse fridge.
 * Improved translations. Thanks to carloshbcabral (updated Portuguese) and Kareolis (updated Russian)!
 
 ## 1.32.3